Upgraded to latest version of MALLET (in some cases, the upgrade needs checking - these are marked by TODOs)
Changed how fanseparser is included. Had originally planned on using non-maven-jar, but I couldn't get it working on my Mac so opted for the hackier system scope. It's a bad way of doing it, but it is a lot more robust. Really, the authors of fanseparser should be encouraged to add it to Maven.

Started tidying up the code, however there is a lot of work that needs doing there. There is also huge amounts of dead and redundant code which could be removed - didn't want to do that here, but it might be worth doing in the future.